+++ CHANGELOG.rst
@@ -2,9 +2,13 @@
 Changelog
 =========
 
+v1.0.4
+~~~~~~
+* Added data_type argument of ChartDataSerie
+
 v1.0.3
 ~~~~~~
-- Added Chart configuration entity
+* *Added Chart configuration entity
 
 v1.0.2
 ~~~~~~